refactor: photo_core のテストを別ファイルへ切り出しgod-module化を抑える(#29)
mod.rs(2054行)とdating.rs(499行)はdoctrineの1ファイル約300行上限を
超えていた。本体ロジックとテストの内容は変更せず、#[path]属性で
mod_tests.rs/dating_tests.rsへ移すことでファイルサイズを縮小する。
